+++ go.mod
@@ -6,8 +6,7 @@ require (
 	github.com/bwmarrin/discordgo v0.20.3
 	github.com/dgraph-io/ristretto v0.0.3
 	github.com/faiface/beep v1.0.2
-	github.com/fatih/color v1.9.0
-	github.com/gorilla/websocket v1.4.2 // indirect
+	github.com/gookit/color v1.4.2
 	github.com/kardianos/osext v0.0.0-20190222173326-2bc1f35cddc0
 	github.com/valyala/fasthttp v1.16.0
 )
